# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.Charger:
    properties:
      address:
        type: string
      id:
        description: gorm.Model
        type: integer
      lat:
        type: number
      lng:
        type: number
      provider_id:
        type: integer
      rates_id:
        type: integer
      status:
        type: string
    type: object
  dto.ChargerRate:
    properties:
      address:
        type: string
      id:
        description: gorm.Model
        type: integer
      lat:
        type: number
      lng:
        type: number
      provider_id:
        type: integer
      rates:
        $ref: '#/definitions/dto.Rates'
      status:
        type: string
    type: object
...
  dto.Rates:
    properties:
      id:
        type: integer
      no_show_penalty_rate:
        type: number
      normal_rate:
        type: number
      penalty_rate:
        type: number
      provider_id:
        type: integer
      status:
        type: string
    type: object
info:
  contact: {}
  title: Provider Service API
  version: "1.0"
paths:
...
  /provider/{provider_id}/charger:
    get:
      consumes:
      - application/json
      description: get Charger by provider id or charger id
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      - description: Provider id
        in: path
        name: provider_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns a list of Charger object
          schema:
            items:
              $ref: '#/definitions/dto.Charger'
            type: array
      summary: Get all Charger
      tags:
      - Charger
    patch:
      consumes:
      - application/json
      description: Update Charger by provider
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      - description: Provider id
        in: path
        name: provider_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns a Charger object
          schema:
            $ref: '#/definitions/dto.Charger'
      summary: Update Charger by provider
      tags:
      - Charger
    post:
      consumes:
      - application/json
      description: create Charger by provider
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      - description: Provider id
        in: path
        name: provider_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns a Charger object
          schema:
            $ref: '#/definitions/dto.Charger'
      summary: Create Charger by provider
      tags:
      - Charger
  /provider/{provider_id}/charger/{charger_id}:
    delete:
      consumes:
      - application/json
      description: Delete Charger by charger id
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      - description: Provider id
        in: path
        name: provider_id
        required: true
        type: integer
      - description: Charger id
        in: path
        name: charger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns a Charger object
          schema:
            $ref: '#/definitions/dto.Charger'
      summary: Delete Charger by charger id
      tags:
      - Charger
  /provider/{provider_id}/chargerandrate:
    get:
      consumes:
      - application/json
      description: get Charger and Rate by provider id
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      - description: Provider id
        in: path
        name: provider_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns a list of Charger object
          schema:
            items:
              $ref: '#/definitions/dto.ChargerRate'
            type: array
      summary: Get Charger and Rate by provider id
      tags:
      - ChargerRate
    patch:
      consumes:
      - application/json
      description: Update Charger and Rate by provider Id
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      - description: Provider id
        in: path
        name: provider_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns a ChargerRate object
          schema:
            $ref: '#/definitions/dto.ChargerRate'
      summary: Update Charger and Rate by provider Id
      tags:
      - ChargerRate
    post:
      consumes:
      - application/json
      description: Create Charger and Rate by provider id
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: returns a ChargerRate object
          schema:
            $ref: '#/definitions/dto.ChargerRate'
      summary: Create Charger and Rate by provider id
      tags:
      - ChargerRate
...
  /provider/charger/{charger_id}:
    get:
      consumes:
      - application/json
      description: get Charger by provider id or charger id
      parameters:
      - description: jwtToken of the user
        in: header
        name: authentication
        type: string
      - description: Charger id
        in: path
        name: charger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: returns a list of Charger object
          schema:
            items:
              $ref: '#/definitions/dto.Charger'
            type: array
      summary: Get all Charger
      tags:
      - Charger
...
schemes:
- http
swagger: "2.0"